Add logic for static registration of plugins
In llvm-tutor we mostly use dynamic plugins. This patch adds
documentation, a bash script and a CI action to demonstrate static
plugin registration as an alternative. The following presentation
provides a great introduction into this subject:

2020 LLVM Developers’ Meeting: S. Guelton
“Building Compiler Extension for LLVM 10.0.1”
https://www.youtube.com/watch?v=PkfTHdoUzT8&ab_channel=LLVM

More specifically, this patch contains a bash script that automates
porting MBASub into an in-tree, static LLVM plugin.